Changed encoder ticks from long to int16.
Since all communication is dealing with 2 byte values, the encoder ticks should also be a 2 byte value rather than Arduino's long (4 bytes). This offers +/- 32768 encoder ticks which is more than enough since one full revolution using a 50:1 gear motor is 12*51.45 = 617.4 counts per revolution. We also assume that we shouldn't be spinning more than a full revolution.

+++ DCMotorPIDTorqueControl/DCMotorPIDTorqueControl.ino
@@ -79,7 +79,7 @@
 
 // Volatile variables for the ISR
 volatile int portState = 0;
-volatile long encoderTicks = 0;
+volatile int16_t encoderTicks = 0;
 volatile bool changedTick = 0;
 
 // Communication
